Move AHI, DART and DOSSB output drivers into own files
