+++ CMakeLists.txt
@@ -24,8 +24,11 @@ OPTION(BUILD_SHARED_LIBS "Build a dynamic wildmidi library" ON)
 OPTION(WANT_PLAYER "Build WildMIDI player in addition to the libraries" ON)
 OPTION(WANT_STATIC "Build static library in addition to dynamic library" OFF)
 CMAKE_DEPENDENT_OPTION(WANT_PLAYERSTATIC "Build a statically linked WildMIDI player" ON "WANT_STATIC;WANT_PLAYER" OFF)
+OPTION(WANT_AMIGA_AHI "Include Amiga AHI support" OFF)
 OPTION(WANT_ALSA "Include ALSA (Advanced Linux Sound Architecture) support" OFF)
+OPTION(WANT_DOSSB "Include DOS SoundBlaster support" OFF)
 OPTION(WANT_OPENAL "Include OpenAL (Cross Platform) support" OFF)
+OPTION(WANT_OS2DART "Include OS/2 DART support" OFF)
 OPTION(WANT_OSS "Include OSS (Open Sound System) support" OFF)
 OPTION(WANT_WIN32_MM "Include Windows Multimedia support" OFF)
 OPTION(WANT_DEVTEST "Build WildMIDI DevTest file to check files" OFF)
@@ -42,6 +45,18 @@ ENDIF ()
 IF ((NOT BUILD_SHARED_LIBS) AND (NOT WANT_STATIC))
     MESSAGE(FATAL_ERROR "Neither dynamic nor static library build is selected.")
 ENDIF ()
+
+IF (WANT_AMIGA_AHI)
+    SET(AUDIODRV_AHI 1)
+ENDIF()
+
+IF (WANT_DOSSB)
+    SET(AUDIODRV_DOSSB 1)
+ENDIF()
+
+IF (WANT_OS2DART)
+    SET(AUDIODRV_OS2DART 1)
+ENDIF()
 
 IF (WANT_WIN32_MM)
     SET(AUDIODRV_WIN32_MM 1)
